fix: altera TotalAmount de double para decimal no CartHeaderViewModel

+++ VShop.Web/Models/CartHeaderViewModel.cs
@@ -6,6 +6,6 @@
         public string UserId { get; set; } = string.Empty;
         public string CouponCode { get; set; } = string.Empty;
 
-        public double TotalAmount { get; set; } = 0.00d;
+        public decimal TotalAmount { get; set; } = 0.00m;
     }
 }
